\name{testUrl}
\alias{testUrl}

\title{
Testing http access to given url
}
\description{
This function checks if wanted text can be read from the given url
}
\usage{
testUrl(url="http://www.survo.fi/muste/PACKAGES",checktxt="Package: muste")
}

\arguments{
  \item{url}{
    Url to be tested
}
 \item{checktxt}{
    First line of text in given url
}
}
\author{
Reijo Sund
}
