\name{automuste}
\alias{automuste}

\title{
Autoupdater for Muste
}
\description{
This function autoupdates package Muste and related packages and autostarts Muste.
}
\usage{
automuste(kysy=FALSE,curl="http://www.survo.fi/muste")
}

\arguments{
  \item{kysy}{
    "Ask" parameter for new.packages and update.packages (FALSE/TRUE/"graphical")
}
 \item{curl}{
    "Contriburl" parameter for package function.
}
}
\author{
Reijo Sund
}
